refactor: extract makeCanvasGameStore factory, persist best score (#308)
catchFruitStore and spaceDefenderStore were copy-paste identical —
same state shape, same 5 actions, only GAME_DURATION differed.

Extract makeCanvasGameStore(name, persistKey, gameDuration) in
lib/stores/utils/canvasGameStore.ts using makePersistStore so best
scores survive page reloads (partially addresses #305).
Reduce each store to 3-line factory call.
